+++ jgclark.Dashboard/CHANGELOG.md
@@ -4,6 +4,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 ## [2.1.3] 2025-01-???
 ### Fixes
 - fixed regression that stopped 'hide checklists' from being honoured in some parts of the display
+- the 'current timeblock' could stop being displayed too soon
 ### Dev notes
 - I realised that some of the logic for what to display and not is spread out in several places, making it difficult to reason with and test. So I've moved a checklist filter out of ItemGrid up to Section level (through its useSection... effect) which has much more logic about how to filter and display this.